fix: corregir errores en control de stock (precios, validacion, error display)
- items.py: fix costo parsing que removia el punto decimal (15864.6 → 158646)
- conteo_stock.py/sugerencias.py: cambiar stock_sistema de int a float
- api.ts: manejar error.detail como array (Pydantic validation errors)
- formatPrecio: agregar 2 decimales en formato argentino

+++ backend/app/routes/conteo_stock.py
@@ -35,7 +35,7 @@ class ProductoTareaCreate(BaseModel):
     cod_item: str
     nombre: str
     precio: float
-    stock_sistema: int
+    stock_sistema: float
 
 
 class TareaConteoCreate(BaseModel):
