+++ backend/app/routes/sugerencias.py
@@ -30,7 +30,7 @@ class ProductoSugerencia(BaseModel):
     cod_item: str
     nombre: str
     precio: float
-    stock_sistema: int
+    stock_sistema: float
 
 
 class SugerenciaCreate(BaseModel):
